A CSS Lexer http://codereview.appspot.com/5265046
This replaces the lexer work in cssparser.js which I plan to gut in a later
CL to build on this lexer.

First though, I plan to replace the CSS handling code in domado.js to use
this to normalize CSS text so that we can use much simpler regexps to
validate CSS properties according to our CSS schema.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4640 360991c0-9f3b-0410-ba87-8b362a00edf4