Supporting class for new CSS Schema->JavaScript converter http://codereview.appspot.com/5372053
This is a supporting class for the rewrite of CssPropertyPatterns which
produces css-defs.js.

It doesn't have anything directly related to CSS, but is instead a bunch of
set manipulation code.

This will be used to compute a partition of a set of CSS literals so that
we can support a larger subset of the CSS schema without blowing up the
amount of code we ship to the client.

Right now, css-defs.js use a simplistic definition of color because <color>
appears in many CSS property definitions, and is a very large production
including keywords for each entry in the X11 color table.

By computing a partition of keywords when compiling css-defs.js we can
support the full color definition because the partition computer realizes
that all the color keywords appear together and so it can create one set
of color name keywords regardless of how many CSS properties use those.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4670 360991c0-9f3b-0410-ba87-8b362a00edf4