early-bind idClass in caja.js and Playground http://codereview.appspot.com/5488053
Right now caja.js and the Playground will call the cajoler
without an idClass parameter, which means that rewriting html ids
is deferred until gadget instantiation time.

I don't see much benefit to doing that, since there isn't an api
for multiply-instantiatiating the same gadget on the same host
page, and there isn't any particular demand for doing so.

This change makes caja.js and the Playground pass an idClass to
the cajoler in some cases, which has a few performance benefits:

1. With early-binding of idClass, the cajoler doesn't have to
generate html-emitter statements to set every html id, which
reduces the amount of js that the cajoler processes, which speeds
up the cajoler.

2. Avoiding all those html-emitter statements also speeds
up client-side rendering.

If in the future we do want to support multiply-instantiating
a gadget without calling cajole each time, then it's easy enough
to add an option for that.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4679 360991c0-9f3b-0410-ba87-8b362a00edf4