faster domado startup: defer creation of TameStyle http://codereview.appspot.com/5489047
TameStyle (and TameComputedStyle) are expensive to define because domado
operates on a large list of style attributes, and defines frozen setters
and getters for each attribute.

This change defers all that work until the first time a TameStyle or
TameComputedStyle is requested.

With this change and the cajaVM.def speedup,
it takes about 850ms to start a guest in Chrome Beta (16.0)
on my machine.

With just the cajaVM.def speedup,
it takes about 1100ms to start a guest.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4688 360991c0-9f3b-0410-ba87-8b362a00edf4