Creates wrapper for certain feral nodes that were not available before. http://codereview.appspot.com/5755043
permitUntaming() associates a useless object with a given DOM object obj so
that untame(obj) doesn't blow up in your face.  When obj is a DOM node, we
often want to recover the feral node; this adds a function untamesToWrapper()
that associates a DOM object with an object containing a "node" property
rather than a useless object.

R=ihab.awad


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4802 360991c0-9f3b-0410-ba87-8b362a00edf4