fix el.contains(null) for jquery UI http://codereview.appspot.com/5756069
jquery UI's jquery.effects.core.js has a method createWrapper
that tries to do:
    el.contains(document.activeElement)
which throws the confusing error message:
    Can't use key lifetime on primitive keys: undefined

The problem is .activeElement is not supported by domado, so the
value is undefined.

el.contains(undefined) works in all browsers, but fails in domado
because it tries to do TameNodeT.coerce(undefined), which throws
that confusing error.

activeElement is not in html4, but it is in html5, and supported
by all browsers.  I'm not implementing that here, because it's
kind of tricky and doesn't seem important.

Instead, this change makes contains(null) and contains(undefined)
return false, matching browser behavior.

Note, the related function compareDocumentPosition(null) does not
have consistent behavior in browsers; Chrome allows it, Firefox
throws an error, so I'm leaving that as an error.

I've also made the TameNodeT.coerce(undefined) error message
slightly less confusing.  Now it will say:
    Specimen does not have the "TameNode" confidence mark
which is still not very comprehensible, but it's closer to the
actual problem.

R=metaweta@gmail.com, jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4803 360991c0-9f3b-0410-ba87-8b362a00edf4