Do URL transformation in client-side CSS rewriter. http://codereview.appspot.com/5841053
This allows the client-side CSS parser involved when
document.write('<style>...</style>') is called to use the existing URL
policy to rewrite background image URLs and the like.

The URL policy is passed an absolute URL, so we resolve the URL using
a base URL contained in html-emitter.js.  Instead of parameterizing
html-emitter further, I just added special handling to its
document.write implementation for the <base> tag.  If it sees a <base>
tag on input, it adjusts its internal base URI appropriately.
If the container wants to specify a base URI, it can document.write a
base tag.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4818 360991c0-9f3b-0410-ba87-8b362a00edf4