don't cache rewriter errors http://codereview.appspot.com/5907044
This fixes issue 1455
http://code.google.com/p/google-caja/issues/detail?id=1455

In some cases, when the rewriter has an error, it emits an error message,
and the cajole fails, but the rewritten js ends up in the job cache.

Since the job cache doesn't cache errors, the next time the same js is
cajoled, the result is fetched from the cache and the cajole succeeds.

R=metaweta, jasvir


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4827 360991c0-9f3b-0410-ba87-8b362a00edf4