add RegExp to list of unmaskable identifiers http://codereview.appspot.com/5916043
The es53 rewriter transforms /foo/ into
  new RegExp.new___('foo');

'RegExp' is not a protected identifier; it's possible for guest
code to assign to RegExp as a local variable.

This is not a full breach because guest code is not able to
set o.new___ or o['new___'], and there isn't an existing object
with a dangerous new___ method.

However, the parital breach makes me nervous, and I don't see a
good reason to allow guest code to assign RegExp.

So this change adds RegExp to the forbidden list.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4834 360991c0-9f3b-0410-ba87-8b362a00edf4