Rework Server side CSS passes so CLASS and ID namespacing occurs in CSSRewriter http://codereview.appspot.com/5975048
This changes CssDynamicExpressionRewriter and CssRewriter so now CssRewriter
is responsible for namespacing IDs and CLASSes in CSS selectors.

It updates the tests to include the rewritten selectors.

I would prefer to implement the CSS selector rewriter to
sanitizecss.js in another CL and reenable the JS side CSS tests that
are now marked @FailureIsAnOption to keep this CL reasonably small.

R=jasvir@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4847 360991c0-9f3b-0410-ba87-8b362a00edf4