remove jsdocs http://codereview.appspot.com/6268044
The ant 'jsdocs' target has been broken for some time and nobody seems
to care.

Also, Caja's classless js coding style is not particularly amenable to
javadoc-style documentation in the first place.

Rather than try to fix jsdoc, this patch deletes jsdoc entirely,
incidentally saving a few seconds in 'ant runtests'.

If we want to resurrect jsdoc sometime in the future, it's easy enough to
pull it out of svn history.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4894 360991c0-9f3b-0410-ba87-8b362a00edf4