update rhino http://codereview.appspot.com/6271046
This updates rhino to an unreleased trunk version,
because the released version has a bug
where the expression Object(Object.prototype)
will cause the interpreter to get stuck in an infinite loop.

I want the bugfix because es53.js has many places that say
  Type(o) === 'Object'
which can be replaced with a more efficient test
  o === Object(o)
but that's unsafe in rhino.

I'm comfortable with using an unreleased rhino because we're
only using rhino for tests.

The change to RhinoAsserts is because rhino added a new
type of string-like object for efficient concatenation.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4901 360991c0-9f3b-0410-ba87-8b362a00edf4