remove some unsafe rendering options http://codereview.appspot.com/6297063
The renderer has a couple flags that mean "give me unsafe output
because I know what I'm doing".  There's no reason to allow that,
and the flags increase testing complexity.

ascii-only means restrict output chars to the range \x01 to \x7f.
When ascii-only is off, we can emit arbitrary unicode chars in an
arbitrary encoding, which is unsafe if the encoding is wrong.

embeddable means emit javascript that can be safely inlined in
html or xml.  When embeddable is off, we can emit javascript with
text like '</script>', which is unsafe.

So this patch makes ascii-only and embeddable always true, and
eliminates the unsafe options.

R=metaweta@gmail.com, ihab.awad@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4919 360991c0-9f3b-0410-ba87-8b362a00edf4