fix sizing of Flash objects http://codereview.appspot.com/6371044
This fixes
http://code.google.com/p/google-caja/issues/detail?id=1333

so that when you embed an untrusted Flash object at an arbitrary size,
it gets scaled instead of clipped.

This is not an ideal fix, because the untrusted object doesn't get to
find out its actual size.

Which means things like Youtube videos will show resized controls;
the Youtube video can't adapt its UI to the actual size.

But I don't think that can be fixed, because setting or querying size
in either direction is crossing a security domain boundary.

R=metaweta@gmail.com


git-svn-id: https://google-caja.googlecode.com/svn/trunk@4953 360991c0-9f3b-0410-ba87-8b362a00edf4